tools/requirements.txt
# Python dependencies for agentic-development-v3 tools
# Install with: pip install -r requirements.txt

# TTS for audiobook generation
edge-tts>=6.1.0

# Image processing for cover upscaling
Pillow>=10.0.0
